--- packages/guard-core-v4/src/CreateTenant/interface.ts
@@ -1,3 +1,0 @@
-export interface Props {
-  onBack?: () => void
-}
