--- packages/guard-core-v4/src/JoinTenant/interface.ts
@@ -1,12 +1,0 @@
-export interface Props {
-  onBack: () => void
-}
-
-export enum JoinTenantStepEnum {
-  InputTenantCode = 'inputTenantCode',
-  InputEnterpriseEmail = 'inputEnterpriseEmail',
-  VerifyEmailCode = 'verifyEmailCode',
-  NoEnterpriseDomain = 'noEnterpriseDomain',
-  JoinSuccess = 'joinSuccess',
-  JoinFailed = 'joinFailed'
-}
