--- packages/guard-core-v4/src/JoinTenant/styles.less
@@ -1,14 +1,0 @@
-.g2-view-join {
-  .authing-ant-input-group-addon  {
-    border-right: none;
-    border-left: 1px solid #E5E6EB;
-    border-top-left-radius: 0;
-    border-bottom-left-radius: 0;
-    border-top-right-radius: 4px;
-    border-bottom-right-radius: 4px;
-  }
-  .authing-ant-form-item-control-input-content {
-    display: flex;
-    justify-content: center;
-  }
-}
