+++ packages/guard-core-v4/src/TenantPortalSelect/InputEmailCode/index.tsx
@@ -1,9 +1,9 @@
 import { useTranslation } from 'react-i18next'
 import { React } from 'shim-react'
-import { VerifyCodeFormItem } from '../MFA/VerifyCodeInput/VerifyCodeFormItem'
-import { VerifyCodeInput } from '../MFA/VerifyCodeInput'
-import { SendCodeBtn } from '../SendCode/SendCodeBtn'
-import { useGuardHttp, useGuardPublicConfig } from '../_utils'
+import { VerifyCodeFormItem } from '../../MFA/VerifyCodeInput/VerifyCodeFormItem'
+import { VerifyCodeInput } from '../../MFA/VerifyCodeInput'
+import { SendCodeBtn } from '../../SendCode/SendCodeBtn'
+import { useGuardHttp, useGuardPublicConfig } from '../../_utils'
 import { EmailScene } from 'authing-js-sdk'
 import { message } from 'shim-antd'
 import { useEffectOnce } from 'react-use'
