+++ packages/guard-core-v4/src/TenantPortalSelect/JoinButton/index.tsx
@@ -1,9 +1,9 @@
 import { React } from 'shim-react'
-import { IconFont } from '../IconFont'
-import { TenantButtonProps } from './interface'
+import { IconFont } from '../../IconFont'
+import { TenantButtonProps } from '../interface'
 import classNames from 'classnames'
 import { useTranslation } from 'react-i18next'
-import './styles.less'
+import '../styles.less'
 
 export const JoinButton: React.FC<TenantButtonProps> = ({
   className,
